requirements.txt
requests
pandas
selenium
webdriver_manager
tkinterdnd2
python-dotenv